Make create-extension-didn't message clearer.
The only error message resulting from a CREATE EXTENSION attempt
that failed (because the library had been loaded before in the
session, therefore LOAD was a no-op) is one about a table that
already exists.

But the name of the table appears in the error message, so the
table may as well be named
"see doc: do CREATE EXTENSION PLJAVA in new session"
which may serve to get the point across.