No more standard charset lookups by name.
Rethink use of java.util.Scanner to slurp data ... it's very
convenient, but part of the reason is it swallows IOExceptions.
If you add the extra lines of code to hang onto the reference
and call its ioException() method to see if something went wrong,
the brevity advantage slips away.