Lexicals.identifierFrom.
Fixed a typo in ISO_UNICODE_IDENTIFIER and added JUnit test (not
necessarily in that order).

This changes Lexicals from an interface to an abstract class. That
should not affect any client code unless it was written to inherit
from Lexicals, which would have been an example of the "constant
interface antipattern" so with any luck will not have been common.