Complete the Javafication of setupFunctionParams.
Its last step: *another* possible adaptation of the return type,
based this time on an explicit return type at the left end of the
AS spec. This can clobber the one done by parseParameters; probably
that's a case that just hasn't come up in practical use. For now, it
still works that way, in the spirit of reimplementing without
substantive behavior changes at the same time.