Use heap_form_tuple, not heap_formtuple.
Both have existed back to PG 8.2, so no need for version conditionals.
The preferred one uses a boolean nulls array instead of a char array
of ' ' or 'n', and the deprecated one goes away in 9.6.

memset(nulls, true, count * sizeof(bool)) is a bit off if bool is ever
wider than a byte (values will be nonzero but not exactly equal to true).
But it's an idiom Tom Lane has used in multiple places:
https://git.postgresql.org/gitweb/?p=postgresql.git;a=blobdiff;f=src/backend/utils/cache/catcache.c;h=d5f237bc;hp=83a12776f;hb=902d1cb35;hpb=492059d
https://git.postgresql.org/gitweb/?p=postgresql.git;a=blobdiff;f=src/backend/utils/adt/arrayfuncs.c;h=3818b181;hp=5304d47fa;hb=cecb6075;hpb=c859308a
which makes it good enough for me.