Handle the widening of FuncCallContext's call_cntr.
This one is a bit stickier, because the value is passed to Java code
that implements ResultSetProvider, an interface in pljava-api. The
signature for assignRowValues can't just be changed.

Instead, for now, a ResultSetProvider is limited to returning INT_MAX
rows, but the limit is checked to ensure a predictable failure.

For the future, in the Java 8 world, the ResultSetProvider interface
could grow a default method largeAssignRowValues that takes a long
row number; the Java 8 default implementation would just signal the
same error, and classes implementing the old interface won't break.
Alas, PL/Java isn't committed to supporting only Java 8 yet.

In fact, the same thing could be done on the quiet: PL/Java could check
whether the class implementing ResultSetProvider happens to have a
largeAssignRowValues method with the right signature, and use it if
present. It would simply have to be covered in documentation, and not
formally declared in the interface until the Java support bar moves to
Java 8.

All of that is future work - this change merely checks and enforces
an INT_MAX limit.