Handle widening of SPITupleTable .alloced and .free.
Not many choices available here. Wider-than-32-bit array indices
in Java seem as far off as ever, so without reimplementing TupleTable
to do something else besides copying the whole megillah into a Java
array, there is no choice but to check and report a suitable error if
the table is too big.