Out with my pessimism; PARALLEL SAFE works fine.
A simple test with a PL/Java function declared PARALLEL SAFE,
and used in a parallel query, did just what it says on the tin,
on the first try.

There may still be cases lurking in which PL/Java's inner workings
would do something a SAFE or RESTRICTED function isn't supposed to
do, but I had imagined there would be problems galore and there aren't,
so the capability may as well be announced and documented.