Suppress eager initialization during pg_upgrade.
During pg_upgrade, dynamic libraries are loaded to confirm
they exist in the new installation, but this happens early, before
pg_upgrade has copied the old schema over. PL/Java should not
react to the absence of its schema by creating one (as it would during
a fresh install), because pg_upgrade is expecting to handle that.

Reported in #117.