Select PG version to build for on mvn command line.
Allow selection of which PostgreSQL version to build for,
on a build host with several versions installed, by passing
the full path of that version's pg_config executable on
with -Dpgsql.pgconfig= on the mvn command line.

This was always possible anyway, by simply prefixing the mvn
command with PATH=desired-version-bin-dir:$PATH, but downstream
package maintainers may have more confidence in this option,
and it parallels the -Dpgconfig= option for the self-extracting
jar at installation time.

Also remove a bit of dead code from pom.xml, from back when
less distinction between build and installation time was being made.