Accommodate PG10 demise of float datetimes.
Because PostgreSQL still has an integer_datetimes GUC (it's just
always true now), the obvious approach would be "do nothing", and
the integer code would always be selected, and the floating code
would be dead. However, the dead code now produces compiler warnings
about implicit declaration of floor() and rint(); presumably, we were
relying on math.h being included in some PostgreSQL header where it
now is not (though I am drawing a blank on where that changed; it's
not in the same upstream commit that desupported float datetimes).
Given that, the cleanest solution here is to conditionally compile
the float-time code only for PG < 10, and one day prune it completely
once the backward-compatibility goal reaches 10.