Register a trigger's transition tables.
The magic that makes tableOld/tableNew appear to exist under
the chosen names, for purposes of queries in SPI, happens when
SPI_register_trigger_data is called, passing the TriggerData
struct that was passed to the handler function, and that struct
has the chosen names and the tuplestores.

PL/Java doesn't do SPI_connect unless and until the called Java
function wants to use the jdbc:default:connection, at which point
the connection is created in Invocation.c and remembered for the
current level. So, that's a fine place to call SPI_register_trigger_data.
The TriggerData struct simply needs to be stashed in the current
Invocation at the time of function entry, so it can be passed to the
SPI function when the time comes.