Do away with never-documented LargeObject code.
In the run-up to releasing 1.5.0, the code to do with large objects
was looked over in a very conservative and cautious way, as fits
code that was never documented or presented as API and didn't have
any test coverage, with little way of knowing whether some client
code somewhere in the wild might have been developed around it.
So, it got carefully updated to 64-bit offsets, and even had a
vulnerability reported because it hadn't been updated to honor
the object access controls added in 9.0.

What all that very conservative analysis failed to notice was,
thanks to a change made back in 77bfc34, coupled with the absence
of test coverage, it has been about eleven years since the last
chance anyone ever had of doing anything useful with a LargeObject
instance. That makes the decision to do away with it much easier.

In all the PostgreSQL versions PL/Java currently supports, all the
functions needed to manipulate large objects are already exposed in
SQL and usable through the JDBC/SPI, without any specific effort
needed in PL/Java. For programming convenience, some later version
(after Java 7 becomes the minimum requirement) could add a simple
utility method to turn an integer LO fd from lo_open into a
SeekableByteChannel that could then be used without further trips
through SPI, but even that would be about optimization and convenience,
not functionality.