Poke migration-management versions for 1.5.1-BETA1.
-packaging/build.xml already makes an update .sql from 1.5.0,
the last released version. No change needed before this release.

A change to past practice in updating the SchemaVariant list
in InstallHelper.java. There has been no change to the schema,
so may as well minimize changes by continuing to call it
the 1.5.0 schema (the first full release to feature it), and
adding 1_5_1_BETA1 only as an alias to it.