Document (lack of) issue #134 regression test.
An obvious place to put it, in examples/Triggers.java, won't really
help, because those all run with SPI already connected and would not
catch the bug.

This is more of a note to the future, to make such a test when there
is more of a regression test infrastructure to run tests from outside
PL/Java itself.