Do casts in ResultSetHandle by column definition list
A function declared to return RECORD or SETOF RECORD is required
by SQL to be followed by a column definition list in any query using it.
If the function returns SETOF RECORD using ResultSetHandle, there is a
possibility that one or more corresponding columns in the definition list
and the ResultSet offered by ResultSetHandle differ in type. Let the type
be cast automatically, just as it would be when using ResultSetProvider
and storing each value into the output ResultSet.

Also convert the SetOfRecordTest example to an annotation-style example
and add an SQLAction that tests it.

Addresses issue #146.