Fix dangling pointer left as GUC boot value.
The clearly-intended-but-forgotten 'static' here never caused
functional trouble in practice, because the value was valid at the only
time it was used. It could, however, cause unexpected bytes to appear
in the boot_val column of the pg_settings view, which in turn appeared
as character-encoding errors in tests of XML processing with
schema_to_xml('pg_catalog', ...) as the source document for testing,
which ended up being how this missing 'static' came to light.