Annotate Connection methods and add more comments.
Add @Override to the methods in SPIConnection that are specified
by the Connection API (to make it easier to spot the ones that
aren't). For those added in JDBC 4.1, the annotation is commented out,
as PL/Java 1.5 still strives to be buildable with Java 6. Once the
back-compatibility horizon is Java 7 or later, those @Overrides can
be uncommented.

It would also be fair to say this has added annotations (or
commented-out annotations) through JDBC 4.2, as it didn't add
any new Connection methods.

The PL/Java-specific and internal methods are now easier to pick out
(they're the ones without @Override annotations), and have some more
extensive comments about what they're doing there. Also moved one
method to be nearer the stuff it pertains to. No code changes (except
to add the specified generic signature on getTypeMap/setTypeMap).
Indentation adjusted in a couple contiguous areas.