Sort implemented from unimplemented JDBC4 methods.
Move those that have some JDBC-allowed behavior (other than
throwing a not-supported exception) into the 'implemented'
category, even when the JDBC-allowed behavior is to do
nothing special. Leave in the 'unimplemented' category just
the methods that ought to do something else, but currently
throw the not-supported exception.

In passing, just provide the natural implementation for
isWrapperFor and unwrap; it's no more trouble than throwing
the exception!

For ResultSetMetaData, pull up their implementations into
AbstractResultSetMetaData, so they don't have to be duplicated
in SPIResultSetMetaData and SyntheticResultSetMetaData.

Some indentation normalized, only in areas touched.