Add VarlenaWrapper.
A VarlenaWrapper.Input presents a readable view of a detoasted
datum residing in native memory. It is constructed from native code,
specifying the memory context to (always copy or) detoast into, and
a ResourceOwner bounding the life of the native copy (which must be
released no later than the specified memory context is reset).

The available() method returns the actual detoasted size.

It would also be easy to support mark/reset, but nothing urgently
needs them at the moment.