Close SAXResult upon endDocument().
To comply with the API spec, SQLXML has to throw an exception if
passed to the database to be stored before being closed. However,
a SAXResult exposes no way for the caller to explicitly close it,
so wrap it in an adapter that will do so after the endDocument()
is processed.