Close StAXResult upon endDocument.
To comply with the API spec, SQLXML has to throw an exception if
passed to the database to be stored before being closed. However,
a StAXResult exposes no way for the caller to explicitly close it
(XMLStreamWriter has a close() method, but its documentation explicitly
forbids it to close the underlying stream!). So, wrap it in an adapter
that does so after the endDocument is processed.

This still leaves the oddity that serializing with StAX does not
reproduce self-closing empty elements.