Choose the Source/Result types to be default.
The spec allows getSource and setResult to be called with null
as the desired source or result class, and to return one that is
chosen with some driver-specific knowledge of what's most natural/
efficient.

It was easy to assume that byte streams would be the natural choice,
as they are what's going on under the hood. However, some testing
shows no time cost and even a slight memory advantage when presenting
input in the form of a SAXSource--at least when presenting it to a
JRE-bundled Transformer, which is likely to be a typical case where
null is passed as the desired class.

On the output side, there is a slight time advantage to StreamResult,
but SAXResult is pretty comparable, and saves the caller the burden
of getting the encoding right, closing the stream at the end, etc.
Interestingly, StAX for output was a little more sparing of memory
than SAX, but slower. It's possible that slowness could turn out to
be something simple (adding a BufferedWriter to the pipeline somewhere,
for example); if such an improvement to StAX timing is possible, it
might make a reasonable default on memory grounds. There's still a
point against StAX on fidelity grounds: when driven by a default JRE
identity transform, it can render empty elements with start and end
tags where they were self-closed in the original.

This commit only changes defaults for when the caller has no preferred
class. It is still possible to request any of the choices.

DOM is quite profligate with memory, and can balloon the Java heap
to something like 20 times the size of the serialized XML value; it
should probably be avoided except when it's exactly what the calling
code is built for.