Unwrappage for StAX.
The XMLStreamReader created here returns the expected sequence of
things with no complaints or surprises when called directly.
However, a JRE default identity transformer using it as a source
will ignore or lose things, even while the 'same' transformer
reading the same sequence of things through the SAXUnwrapFilter
reproduces them faithfully. Odd, but for now it has to be enough
that the filter itself behaves as expected.