Test imperfectly for DOCUMENT/CONTENT in Readable.
The current stored form of XML in PostgreSQL does not save any
indication of whether the XML is in DOCUMENT or CONTENT form (which,
to determine correctly, requires fully reparsing it). Java's bundled
XML parsers do not accept the CONTENT form, requiring a synthetic
document root element to be wrapped around the stream content first,
then filtered out of the parse result. This, of course, is not necessary
if the input is in DOCUMENT form, but is harmless unless the input is in
DOCUMENT form with a DTD, in which case such wrapping must be avoided.

An imperfect test is added here that uses an input stream supporting
mark/reset, and pulls a few initial parse events before resetting the
stream for the application to parse. Those initial events only need to
be pulled until a DTD is parsed (meaning the input is definitely
DOCUMENT), or a disallowed-in-prolog production is parsed (meaning it
definitely isn't), or an element-start is parsed (meaning it could be
either, barring a full reparse to settle the question). The last two
cases can both be handled by the wrapping technique, so don't need to be
distinguished further.

The JDBC spec is mute on how SQLXML should handle DOCUMENT vs.
CONTENT. For the forms of Source that expose parse results,
wrapping and filtering is effective, but it would seem a POLA
violation to include synthetic wrapper elements via
getBinaryStream, getCharacterStream, or getString (or getSource
with StreamSource.class). So, for those cases, no wrapping.
The application code will need to know what it's doing if the
XML value might be of CONTENT form.

Because the JRE-supplied SequenceInputStream doesn't support mark/reset,
adds a (possibly general-purpose) MarkableSequenceInputStream that does,
as long as its constituent input streams do.