Introduce DualState class.
Yes, it is *another* pattern for objects that have Java and native state.
For now, the existing ones are being left alone. The existing patterns
do not seem to include anything for objects that are spec'd to live for
the whole transaction (as JDBC specs SQLXML to do). This pattern hooks
into PostgreSQL ResourceOwner mechanisms (and there is a TopTransaction
ResourceOwner) in order to be usable for that. If nothing blows up, it
may be possible to migrate other existing patterns to it later.