Implement SQLXML.setResult(DOMResult.class)
... the one flavor of Result originally left for later (= now).

It turns out that, alone among the setResult flavors, a DOMResult
does object to being used as the result of a transformation that
produces a document fragment (CONTENT, in SQL/XMLese). That's because,
unless setNode() is called first with a DocumentFragment node, the
transform will default to installing a Document node there instead.

Add an example presetting the DOMResult with an empty DocumentFragment.