Add an example doing "XML echo" at a lower level.
While the simple use of Transformer.transform(src, rlt) in
echoXMLParameter() is easy to read and covers a lot of cases without
a lot of coding, lowLevelXMLEcho() shows that, for some cases,
it is not hard to do without the Transformer. Comparisons of time and
memory demands can be made.