Don't let stream get closed in useWrappingElement.
The JRE XMLStreamReader seems to end up closing it, if it happens
to parse to the end, under some circumstances only, such as if
there's nothing (after the possible decl) but whitespace. I suspect
a bug (XMLDocumentScannerImpl$PrologDriver seems to use
XMLEntityScanner.skipSpaces to do that work, and maybe an /entity/
scanner would assume it can close its stream when it's done); in any
case, make sure that doesn't actually close the original stream.