Allow returning/storing a 'readable' SQLXML.
Relax the distinction between 'readable' and 'writable' SQLXML
instances just enough to allow a readable one passed in (or, later,
retrieved from a ResultSet) to be directly returned (or, later,
passed as a parameter or into a ResultSet), as long as it is still
readable, i.e., has not had any of the reading methods called by Java.

It can be used that way exactly once, as the native code retakes
ownership of it at that point.