SQLXML in/out of PreparedStatements / ResultSets.
For PreparedStatement.setObject(col, x, Types.SQLXML) to work,
it's sufficient to mention the type in Oid.c. Getting the type
back from the ResultSet already works, thanks to the
Type_registerType() already done in SQLXMLImpl.c. However, the
test in the getObject(col, SQLXML.class) needed fixing, as it was
testing exact class equality, not whether it got an implementation
of the desired interface. (In passing, reorder those methods to make
clearer what JDBC revision brought them in.)

The setObject overload without the explicit Types.SQLXML will not
work yet, and simply adding addType(SQLXML.class,Types.SQLXML)
in SPIConnection is not sufficient ... because that uses a hash
table and therefore also depends on exact class equality; the full
grody names of the implementing classes would have to be added.
Or the mechanism would have to be fixed.

Wondering why the out.updateObject(col, x) works, without giving
an explicit type? That result set is already typed (by the column
definition list you have to include in the query calling this
RECORD-typed function). That method succeeding is /almost/ enough
to make a subsequent ps.setObject(col, x) work right (because in
succeeding, it primed the s_class2typeId cache in Oid); how would
that be for spooky action at a distance? But that doesn't quite
happen, because that cache is also a strict-class-equality hash map.

At this point, implementing the various (get,set,update}SQLXML methods
will be an easy formality, as they can be implemented in terms of the
methods taking explicit types, which now work.

Alas, this current behavior won't do for a 1.5.x release, because it
would actually start producing SQLXML objects by default where old
code may be expecting String. Before merging, this will need a way
to suppress that automatic mapping but still allow the explicit kind.
A subsequent major release can then go the rest of the way.