A largely-working XMLQUERY using s9api.
Known limitations:
Maps all INTERVALs to xs:duration, rather than to
yearMonthDuration or dayTimeDuration as called for in the spec
(PL/Java ResultSetMetaData not exposing adequate information
from the typmod).

Maps all NUMERIC / DECIMAL to xs:decimal, not to xs:integer
when scale is zero as called for in the spec (PL/Java
ResultSetMetaData returning hardcoded value for scale).

Surely needs more work around time zones for types that may
or may not have them.