Let annotations give row-type parameters defaults.
The @SQLType annotation can now supply an array of strings as
defaultValue for a parameter that has a row type (either the RECORD
type, whose structure is unknown, or a named row type).

In the case of RECORD, the only default that will be accepted is
{} (the array of no strings). While a bit limiting, that is still
just the ticket when a RECORD-typed parameter is being used to
supply a function with an optional, arbitrary sequence of named, typed
parameters.

In the case of a named row type, the default should be an array of
as many strings as the components of the type, and the strings should
be castable to the corresponding component types. The DDR generator
has no way to check that at compile time, but PostgreSQL will report
the error at jar install time if there is a mismatch.