Tidy, improve internal interfaces.
Factor the main works of XMLQUERY into an xmlquery_internal method
that is easier to call, and move the result serialization as CONTENT
back out into xq_ret_content (where the form of result is enforced).
Also create a Binding class to abstract the bindings of context item
and parameters away from ResultSet specifics; BindingsFromResultSet
is now just one implementation.

There turns out to be a handy isValidNCName() provided in Saxon.