A first cut at XMLTABLE.
Correctly returns atomic types matching atomic values passed in,
except times and timestamps aren't right yet, to the great surprise
of no one in particular.

Is limited, so far, to output columns of non-XML types, where each
column expression produces an atomic value that's an instance of the
atomic type the rules determine from the SQL output column type. The
automatic casting of some other atomic type result value AV to the
value BV of the determined type isn't implemented yet (but that is no
blocker; XQuery "cast as" can be written into the column expression
to achieve exactly that effect).

This is perhaps most noticeable with the output column type bytea,
for which the rules always determine an XQuery type of xs:hexBinary.
A binary value in the form xs:base64Binary can't be retrieved as a
bytea, without writing "cast as xs:hexBinary" in the column expression.
That limitation will go away when the AV-to-BV casting rules are in place.