Get negative intervals working.
Both XML Schema and PostgreSQL supposedly accept the ISO 8601
PdddYddMddDTddHddMdd.dddS form, but they could hardly disagree
more on what to do if the value is negative. XML Schema wants
-PdddYddMddDTddHddMdd.dddS which PG roundly rejects, wanting
to see instead P-dddY-ddM-ddDT-ddH-ddM-dd.dddS which is anathema
to XS. For this datatype there is no real help to be had from
JDBC or Saxon methods, so it boils down to shuffling the '-'
characters around in the lexical string.