Correct errors in Timestamp (issue #155).
Review the arithmetic to make sure 'floored' / and % are always done
to preserve the constraint that the 'nanos' component of the Java
Timestamp class is always nonnegative, and to avoid double-counting
the part of the fractional second that is represented in both the
millis and the nanos components.

Add SQLActions in the TypeRoundTripper to test this behavior.

In passing, add conditionals for access in MSVC to the session_timezone
GUC, for those PG versions where it has become possible, and avoid the
work of computing tz in coerceDatum unless it is needed.