This fixes a problem but not the problem.
The name passed to getTypeElement is expected to be a canonical name,
but that isn't what Class.getName() returns (or there wouldn't be a
Class.getCanonicalName()).

However, the problem seems to be deeper; arrays are not listed among the
things TypeElements exist to represent. It may be necessary to drill down
to the component type and then construct the TypeMirror using
getArrayType().