Adapt the Parameters example to annotation style.
This example checks many of the mappings between SQL and Java
parameter and return types, so, as an annotation example, it
also exercises those mappings in the SQL generator. It is
sufficient to expose issue #157 (indeed, with this commit only,
compilation will fail on the inability to map byte[] to bytea).

The former, hand-crafted deployment descriptor did not contain
a declaration for every public static method in the Parameters
class. Those that were not declared then have been left
unannotated now.

The SQL generator includes parameter names in the function
declarations it emits. That makes it possible to use named-
parameter notation when calling the functions. However, it
has a consequence that parameter names should not be SQL
type keywords, as parameter names are optional in declarations
so if the name could be a type, PostgreSQL assumes it is, then
fails to parse the actual type that follows it. A few Java method
parameter names, accordingly, have been changed here.