Don't repeat yourself with UDT mappings.
It was never satisfying to define a UDT in Java through annotations
and still have the DDR generator give "no mapping to an SQL type" errors
for functions accepting or returning that type. It was easy enough (if
tedious) to work around by adding explicit type= or @SQLType annotations,
but the DDR generator had all the necessary information to figure that out
without such manual help. Now it does, eliminating the tedium
illustrated in jcflack/pljava-udt-type-extension@bcb5734.

To accommodate mappings added from the source being compiled, the
protoMappings collection is now of TypeMirror instances rather than
of Class instances.

Still future work: also add implied ordering dependencies for uses of
the new type (such as functions that have it as parameter or return
types and have to follow the type declaration, as illustrated in
jcflack/pljava-udt-type-extension@225ef5c, or that are named in it as,
for example, typmod in/out functions, and have to precede it).