Make sure InstallHelper_groundwork has a snapshot.
The check in pg_plan_query() was added for 8.4 (backpatched
to 8.3.6.and 8.2.12), and can prevent loading in some cases
(seen in 9.1 and earlier), at least when doing the 'guided'
sort of load with interactive setting of GUCs.

In passing, fix a plain old typo in Type that has been there
for a while, but only is exposed when building for 8.2.