Preserve LOAD usability over supported PG versions
In PG versions where CREATE EXTENSION is available, it will surely be
the most common installation method, but the method using LOAD should
not be allowed to bitrot. It can still be preferable, whether for quick
testing purposes, or cases where write access isn't available to the PG
installation's libdir and sharedir to install the PL/Java files there.

In both the most recent supported PG versions, 10 and later, and the
very earliest, 8.2, there are slight differences in the path of things
to dereference while finding the LoadStmt to get the library name.
In 10 and later, there can be an extra PlannedStmt to unwrap on the way,
and 8.2 had a list of query trees on the Portal instead of Stmts.