Emit REMOVE section of DDR in right order (#163).
The implicitly-added provides/requires relationships between implementor
tags and the snippets that test the environment to activate or suppress
them need to be treated specially. Unlike other dependency
relationships, which have their sense reversed between the INSTALL and
REMOVE action groups, these implicit ones need to keep the same sense.
It always works better to test a condition before needing to use the
result.

The old augmentRequires method, which created the implied dependencies
by outright adding them to the snippet's other requires, is now gone;
the implied relationship is simply handled within the ordering code.
Two DAGs are created, one for install and one for remove, and an order
is found independently for each.

It's also necessary to use the deployStrings of tag-testing snippets
at undeploy time, at least in the case they don't have explicit
undeployStrings supplied, which is the usual case, because the
conditions to be tested are the same deploying and undeploying.