Plant a DualState_cleanEnqueuedInstances call.
Where to plant such calls may involve some arbitrary and strategic
choices, but the one currently planted in Invocation_popInvocation
is surely not enough in the case a PL/Java function loops through
a lot of data, possibly creating and discarding DualState objects,
without returning to PostgreSQL.

This seems a good place to plant another. It is not called for every
row processed, but for every fetchSize rows, defaulting (in
SPIStatement) to 1000. Perhaps that might still be too seldom for
code that churns a lot of DualState objects. But it's fairly certain
not to be too often and impose an onerous overhead.