Add a Verifier for SQLXMLImpl.Writable.
Because the stored form in PostgreSQL of the XML data type is just its
character serialization, of all the methods offered by java.sql.SQLXML
for reading/writing the content, the ones based on strings, character
streams, and byte streams are the quickest ones to get working in a lazy
sort of way.

But the lazy way isn't appropriate for production, at least for the
writing methods, because type safety precludes letting client code write
arbitrary bytes into a varlena that PostgreSQL believes to be of XML
type. So the quick, lazy methods ultimately become the ones that need
the complexity of a Verifier making sure what is written is XML.

The other writing methods, where the serialization is done via Java XML
APIs, are left with the NoOp verifier, trusting the APIs to produce XML.