Add VarlenaWrapper.Verifier.
A Verifier can be set on a VarlenaWrapper.Output to read along as bytes
are written to the varlena, and throw an exception if they do not make a
well-formed representation of the expected type. This is not necessary
if the PL/Java code for a given type retains complete control over what
is written to the varlena, but is, if an API requires exposing the
VarlenaWrapper.Output more or less directly to client code that could
write arbitrary content on it.

The code that implements a specific data type over varlena can create a
subclass of Verifier.Base, whose verify(InputStream) method simply reads
the whole stream and returns successfully if nothing is wrong, otherwise
throws an exception. VarlenaWrapper.Output will take care of submitting
that verify method to a thread pool, feeding it buffers of written data
as they are filled, and collecting its success or exception status when
the stream is closed (or canceling the task when appropriate).