Let MarkableSequenceInputStream draw from a queue.
Change the internal form of the sequence of constituent streams from an
array to a ListIterator backed by a LinkedList, and add a constructor
that accepts a BlockingQueue of constituent streams. If these
constituent streams are markable/resettable, the combined stream
still is also.

This can then serve as an efficient markable input stream for reading
along concurrently with content generated in a series of buffers that
are just handed over here as they are filled.