Avoid 'bouncing' SQLXML unchecked between types.
Commit a73abee made it possible to take a readable SQLXML, as might
come in as a function parameter, and use it directly as you could a
writable one, returning it from the function or stuffing it into a
query parameter.

Commit f152658 made it possible to use this API even in PostgreSQL
instances that were not built with the XML type, by allowing a
function declaration to explicitly type an SQLXML parameter or return
type as 'text'.

In combination, those two features could undermine type safety: a
function like:

  @Function
  public static SQLXML badcast(@SQLType("text") SQLXML sx)
  {
    return sx;
  }

could serve as an "unchecked cast" returning non-XML text content
as a PostgreSQL XML type.

Solution: remember the PostgreSQL type Oid of the source when creating
a readable SQLXML instance. If it is later used directly in the Java->PG
direction, and the target PG Oid is different, run the verifier over the
content first.

That is, of course, slower than not doing it, but interestingly, the
speed of the Java XML parser compares respectably to a direct PG
CAST (text AS xml). I am seeing it pull ahead of the PG cast right
around 32 kB of serialized XML, ultimately beating PG by a factor of
two or better at 1 or 2 MB sizes. PG, of course, wins hands down on
tiny values, where the invocation overhead predominates.