Lazy detoast for 9.5 ... 8.2.
Before 9.6, there was no GetOldestSnapshot(). There may have been some
more roundabout way to dig up a suitable snapshot to protect the toast
rows, but instead, just give up on parking toast pointers and park the
fetched-but-not-uncompressed contents instead. With XML that can still
be an easy factor of 20 or so smaller.

In theory, pre-8.4, it might be possible to park toast pointers again,
because back then they were safe from vacuuming as long as the overall
transaction remains open. But for versions that old, who has the time?