Implement JDBC 4.1 ResultSet.getObject.
This is the version that takes a Class to determine the class of result
to return. To preserve back-compatibility for now, the underlying
implementation method is not constrained to return an object of the
requested class, and accepts null for that argument, and all the
pre-existing methods call it that way, to get the previous behavior.

Per a comment in Type.h, it's always been the idea that TypeObtainer
methods return singletons (or return from a small number of allocated
instances, if there need to be instances with a few different typeIds).
That doesn't seem to have been followed always, but it grows in
importance with this change; certainly any new TypeClass that is added
and expected to be used in a JDBC 4.1/4.2ish way should be careful not
to allocate a new Type instance every time it gets mentioned.