Include return types in most DDR method signatures
The 'AS' syntax for PL/Java function declarations already accepts
explicit return types as well as the parameter types, but the DDR
generator has not been emitting them. That means the magic doesn't
happen when, say, you have a Java method returning one of the JSR 310
date/time types instead of the expected ones from java.sql. For PL/Java
to notice the automatic coercion needed there, it has to see that
the Java return type is not the one that naturally corresponds to
the SQL type ... so the Java return type needs to be included
in the signature.

Avoid, however, including the return type in the case of trigger
functions, or those that return sets or composite types. Those
already get special treatment in the function parsing code, and
the explicit return types would get in the way.