Introduce TypeBridge.
For now, in the 1.5.x series, this is a simple stopgap so that the few
places in PL/Java where an object type can be passed to PostgreSQL
(SingleRowWriter, TriggerResultSet, SQLOutputToTuple, PreparedStatement)
are able to pass an object that isn't of the class expected by default,
and have the right native conversion get selected. All of those sites
currently work by some variant of putting supplied objects into an
Object array or list later passed to the native code, and when an object
will not be of the expected class, what is stored in the array should be
a TypeBridge.Holder for it.

This may be a temporary class that goes away entirely in a future major
release of PL/Java that revamps how type mappings are determined.  Or,
it may evolve and take on greater responsibility in a revamped scheme:
type mapping information is, at present, diffused and duplicated a lot
of places in PL/Java, and bringing it into one place would not be a bad
thing.