Put TypeBridge in play for SPIPreparedStatement.
In passing, complete the implementation of the 3-argument form of
setNull (where the third argument is a PostgreSQL type name), which is
the only method on a PreparedStatement by which a parameter type can be
set to a specific PostgreSQL type (as opposed to a generic JDBC type
with some default mapping to a PostgreSQL type).

However, the implementation is still limited, in that any subsequent
non-null assignment will reset the PostgreSQL type to the default
mapping, if done through one of the pre-JDBC 4.2 setter methods, or
setObject with one of the pre-JDBC 4.2 accepted object classes. That is
to avoid a material behavior change in a minor release.

By contrast, if setObject is called with an object of one of the
newly-accepted classes (which had no prior behavior to try to match),
any already-assigned PostgreSQL type (as by setNull) will be respected,
and the setObject treated as a coercion to it (which can entail an
exception if no such coercion is available).

That new behavior ought to be the way everything behaves in some future
major release, which should use the PG 9.0 improved SPI to determine
the parameters' PostgreSQL types from what PG's type inference came up
with, and treat all the setter methods as coercions to the determined
types.